requirements.txt
numpy==1.23.5
pandas>=1.5.0
yfinance>=0.2.0
PyYAML>=6.0